refactor: event based loop instead of recursive future
